+++ docs/changelog.md
@@ -19,6 +19,15 @@ All notable changes to this project will be documented below.
 
 * v0.1dev: spectral = **geospatial.read_geotif**(*filename, bands="B,G,R", cropto=None*)
 
+#### geospatial.shapes.grid
+
+* v0.1dev: cells = **geospatial.shapes.grid**(*img, field_corners_path, out_path, num_ranges, num_columns,
+         range_length, row_length, num_rows=1, range_spacing=0, column_spacing=0*)
+
+#### geospatial.shapes.flexible
+
+* v0.1dev: cells = **geospatial.shapes.flexible**(*img, field_corners_path, plot_geojson_path, out_path, range_length, row_length, num_rows=1*)
+
 #### geospatial.transform_points
 
 * v0.1dev: coord = **geospatial.transform_points**(*img, geojson*)